Refactor Token Verification
