+++ plugins/pom.xml
@@ -23,7 +23,7 @@
     implementations do not have to do this; they can redeclare those deps. -->
     <groupId>com.google.fhir.gateway</groupId>
     <artifactId>fhir-gateway</artifactId>
-    <version>0.1.35</version>
+    <version>0.1.36</version>
   </parent>
 
   <artifactId>plugins</artifactId>
