+++ pom.xml
@@ -27,7 +27,7 @@
 
   <groupId>com.google.fhir.gateway</groupId>
   <artifactId>fhir-gateway</artifactId>
-  <version>0.1.35</version>
+  <version>0.1.36</version>
   <packaging>pom</packaging>
 
   <name>FHIR Information Gateway</name>
