+++ server/pom.xml
@@ -21,7 +21,7 @@
   <parent>
     <groupId>com.google.fhir.gateway</groupId>
     <artifactId>fhir-gateway</artifactId>
-    <version>0.1.35</version>
+    <version>0.1.36</version>
   </parent>
 
   <artifactId>server</artifactId>
